Restructure marlowe-test to avoid library depending on tasty and plutus scripts
