+++ marlowe-test/marlowe-test.cabal
@@ -1,4 +1,4 @@
-cabal-version: 2.2
+cabal-version: 3.0
 name:          marlowe-test
 version:       0.1.3.0
 license:       Apache-2.0
@@ -118,9 +118,53 @@ library
   exposed-modules:
     Data.Jsonable
     Spec.Marlowe.Common
+    Spec.Marlowe.Plutus.Arbitrary
+    Spec.Marlowe.Reference
+    Spec.Marlowe.Semantics.Arbitrary
+    Spec.Marlowe.Semantics.AssocMap
+    Spec.Marlowe.Semantics.Golden
+    Spec.Marlowe.Semantics.Golden.Escrow
+    Spec.Marlowe.Semantics.Golden.Negative
+    Spec.Marlowe.Semantics.Golden.Pangram
+    Spec.Marlowe.Semantics.Golden.Swap
+    Spec.Marlowe.Semantics.Golden.Trivial
+    Spec.Marlowe.Semantics.Golden.ZeroCouponBond
+    Spec.Marlowe.Semantics.Next.Arbitrary
+    Spec.Marlowe.Semantics.Orphans
+    Spec.Marlowe.Semantics.Path
+    Spec.Marlowe.Semantics.Util
+    Spec.Marlowe.Service
+    Spec.Marlowe.Service.Random
+    Spec.Marlowe.Service.Serialization
+    Spec.Marlowe.Service.Types
+
+  other-modules:   Paths_marlowe_test
+  build-depends:
+    , aeson >=2 && <3
+    , base >=4.9 && <5
+    , bytestring >=0.10.12 && <0.12
+    , containers ^>=0.6.5
+    , directory ^>=1.3
+    , filepath ^>=1.4
+    , marlowe-cardano ==0.1.2.0
+    , mtl >=2.2 && <3
+    , plutus-ledger-api ==1.0.0.1
+    , plutus-script-utils ==1.2.0.0
+    , plutus-tx ==1.0.0.0
+    , QuickCheck >=2.14 && <3
+    , tasty-quickcheck ^>=0.10
+
+  if flag(trace-plutus)
+    cpp-options: -DTRACE_PLUTUS
+
+test-suite marlowe-test
+  import:         lang
+  hs-source-dirs: test
+  type:           exitcode-stdio-1.0
+  main-is:        Spec.hs
+  other-modules:
     Spec.Marlowe.Marlowe
     Spec.Marlowe.Plutus
-    Spec.Marlowe.Plutus.Arbitrary
     Spec.Marlowe.Plutus.AssocMap
     Spec.Marlowe.Plutus.Lens
     Spec.Marlowe.Plutus.MList
@@ -131,23 +175,12 @@ library
     Spec.Marlowe.Plutus.Transaction
     Spec.Marlowe.Plutus.Types
     Spec.Marlowe.Plutus.Value
-    Spec.Marlowe.Reference
     Spec.Marlowe.Semantics
-    Spec.Marlowe.Semantics.Arbitrary
-    Spec.Marlowe.Semantics.AssocMap
     Spec.Marlowe.Semantics.Compute
     Spec.Marlowe.Semantics.Entropy
     Spec.Marlowe.Semantics.Functions
-    Spec.Marlowe.Semantics.Golden
-    Spec.Marlowe.Semantics.Golden.Escrow
-    Spec.Marlowe.Semantics.Golden.Negative
-    Spec.Marlowe.Semantics.Golden.Pangram
-    Spec.Marlowe.Semantics.Golden.Swap
-    Spec.Marlowe.Semantics.Golden.Trivial
-    Spec.Marlowe.Semantics.Golden.ZeroCouponBond
-    Spec.Marlowe.Semantics.Merkle
+    Spec.Marlowe.Semantics.Golden.Tests
     Spec.Marlowe.Semantics.Next
-    Spec.Marlowe.Semantics.Next.Arbitrary
     Spec.Marlowe.Semantics.Next.Common.Isomorphism
     Spec.Marlowe.Semantics.Next.Common.QuickCheck
     Spec.Marlowe.Semantics.Next.Common.Tuple
@@ -157,19 +190,12 @@ library
     Spec.Marlowe.Semantics.Next.Contract.When.Choice
     Spec.Marlowe.Semantics.Next.Contract.When.Deposit
     Spec.Marlowe.Semantics.Next.Contract.When.Notify
-    Spec.Marlowe.Semantics.Orphans
-    Spec.Marlowe.Semantics.Path
-    Spec.Marlowe.Semantics.Util
+    Spec.Marlowe.Semantics.Path.Tests
     Spec.Marlowe.Serialization
     Spec.Marlowe.Serialization.CoreJson
     Spec.Marlowe.Serialization.ExtendedJson
-    Spec.Marlowe.Service
     Spec.Marlowe.Service.Isabelle
-    Spec.Marlowe.Service.Random
-    Spec.Marlowe.Service.Serialization
-    Spec.Marlowe.Service.Types
 
-  other-modules:   Paths_marlowe_test
   build-depends:
     , aeson >=2 && <3
     , base >=4.9 && <5
@@ -177,13 +203,13 @@ library
     , cardano-api ==1.35.4
     , containers ^>=0.6.5
     , data-default ^>=0.7.1
-    , directory ^>=1.3
     , filepath ^>=1.4
     , hint ^>=0.9
     , ilist ==0.4.0.1
     , lens >=5.2 && <6
-    , marlowe-cardano ==0.1.2.0
+    , marlowe-cardano:{marlowe-cardano, plutus-scripts} ==0.1.2.0
     , marlowe-spec-test ==0.1.0.0
+    , marlowe-test
     , mtl >=2.2 && <3
     , plutus-ledger ^>=1.2
     , plutus-ledger-api ==1.0.0.1
@@ -199,20 +225,6 @@ library
     , text >=1.2.4 && <2
     , these >=1.1 && <2
     , type-iso ==1.0.1.0
-
-  if flag(trace-plutus)
-    cpp-options: -DTRACE_PLUTUS
-
-test-suite marlowe-test
-  import:         lang
-  hs-source-dirs: test
-  type:           exitcode-stdio-1.0
-  main-is:        Spec.hs
-  build-depends:
-    , base >=4.9 && <5
-    , marlowe-test
-    , tasty
-    , tasty-quickcheck
 
 executable marlowe-spec-client
   import:         lang
