+++ marlowe-test/src/Spec/Marlowe/Semantics/Golden.hs
@@ -1,49 +1,34 @@
------------------------------------------------------------------------------
+{-# LANGUAGE RecordWildCards #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+
+-- | Golden tests of contract execution.
 --
 -- Module      :  $Headers
 -- License     :  Apache 2.0
 --
 -- Stability   :  Experimental
 -- Portability :  Portable
---
------------------------------------------------------------------------------
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE RecordWildCards #-}
-{-# OPTIONS_GHC -fno-warn-orphans #-}
-
--- | Golden tests of contract execution.
 module Spec.Marlowe.Semantics.Golden (
   -- * Types
   GoldenCase,
   GoldenTransaction,
-
-  -- * Testing
-  tests,
 
   -- * Reference contracts
   goldenContracts,
   goldenTransactions,
 ) where
 
-import Data.List (intercalate)
-import Language.Marlowe.Core.V1.Semantics (TransactionInput, TransactionOutput (..), computeTransaction, playTrace)
+import Language.Marlowe.Core.V1.Semantics (TransactionInput, TransactionOutput (..), computeTransaction)
 import Language.Marlowe.Core.V1.Semantics.Types (Contract, State (State))
-import Language.Marlowe.FindInputs (getAllInputs)
 import Plutus.V2.Ledger.Api (POSIXTime)
-import Test.Tasty (TestTree, testGroup)
-import Test.Tasty.HUnit (assertBool, testCase)
 
 import qualified PlutusTx.AssocMap as AM (empty)
-import qualified Spec.Marlowe.Semantics.Golden.Escrow as Escrow (contract, invalids, valids)
-import qualified Spec.Marlowe.Semantics.Golden.Negative as Negative (contract, invalids, valids)
-import qualified Spec.Marlowe.Semantics.Golden.Pangram as Pangram (contract, invalids, valids)
-import qualified Spec.Marlowe.Semantics.Golden.Swap as Swap (contract, invalids, valids)
-import qualified Spec.Marlowe.Semantics.Golden.Trivial as Trivial (contract, invalids, valids)
-import qualified Spec.Marlowe.Semantics.Golden.ZeroCouponBond as ZCB (contract, invalids, valids)
-
--- | Set to `True` to print the paths through the golden contracts.
-_GENERATE_TEST_CASES_ :: Bool
-_GENERATE_TEST_CASES_ = False
+import qualified Spec.Marlowe.Semantics.Golden.Escrow as Escrow (contract, valids)
+import qualified Spec.Marlowe.Semantics.Golden.Negative as Negative (contract, valids)
+import qualified Spec.Marlowe.Semantics.Golden.Pangram as Pangram (contract, valids)
+import qualified Spec.Marlowe.Semantics.Golden.Swap as Swap (contract, valids)
+import qualified Spec.Marlowe.Semantics.Golden.Trivial as Trivial (contract, valids)
+import qualified Spec.Marlowe.Semantics.Golden.ZeroCouponBond as ZCB (contract, valids)
 
 -- | The golden contracts.
 goldenContracts :: [Contract]
@@ -56,80 +41,8 @@ goldenContracts =
   -- Note that Pangram is omitted because `getAllInputs` takes 30 minutes for it.
   ]
 
--- | Run tests.
-tests :: TestTree
-tests =
-  testGroup
-    "Golden"
-    [ testGolden "Escrow" _GENERATE_TEST_CASES_ Escrow.contract Escrow.valids Escrow.invalids
-    , testGolden "Pangram" _GENERATE_TEST_CASES_ Pangram.contract Pangram.valids Pangram.invalids
-    , testGolden "Swap" _GENERATE_TEST_CASES_ Swap.contract Swap.valids Swap.invalids
-    , testGolden "Trivial" _GENERATE_TEST_CASES_ Trivial.contract Trivial.valids Trivial.invalids
-    , testGolden "Zero Coupon Bond" _GENERATE_TEST_CASES_ ZCB.contract ZCB.valids ZCB.invalids
-    , testGolden "Negative" _GENERATE_TEST_CASES_ Negative.contract Negative.valids Negative.invalids
-    ]
-
--- | Run a golden test.
-testGolden
-  :: String
-  -- ^ Name of the test.
-  -> Bool
-  -- ^ Whether to print the valid test cases instead of doing testing.
-  -> Contract
-  -- ^ The contract.
-  -> [GoldenCase]
-  -- ^ Cases where the contract should succeed.
-  -> [GoldenCase]
-  -- ^ Cases where the contract should fail.
-  -> TestTree
-  -- ^ The tests.
-testGolden name printIt contract valids invalids =
-  testGroup name $
-    if printIt
-      then [testCase "Generate valid inputs" $ printValids contract]
-      else
-        [ testCase "Valid inputs" $ testValid contract valids
-        , testCase "Invalid inputs" $ testInvalid contract invalids
-        ]
-
 -- | A golden test case.
 type GoldenCase = (POSIXTime, [TransactionInput], TransactionOutput)
-
--- | Use static analysis to generate and print successful golden test cases for a contract. This may take many minutes to execute.
-printValids :: Contract -> IO ()
-printValids contract =
-  do
-    Right is <- getAllInputs contract
-    sequence_
-      [ case playTrace t contract i of
-        Error{} -> pure ()
-        o -> putStrLn $ intercalate "\t" [show t, show i, show o]
-      | (t, i) <- is
-      ]
-
--- | Test successful execution for the Pangram contract.
-testValid :: Contract -> [GoldenCase] -> IO ()
-testValid = testValidity True
-
--- | Test erroneous transactions for the Pangram contract.
-testInvalid :: Contract -> [GoldenCase] -> IO ()
-testInvalid = testValidity False
-
--- | Test erroneous transactions for a contract.
-testValidity
-  :: Bool
-  -- ^ Whether the test cases should pass.
-  -> Contract
-  -- ^ The contract.
-  -> [GoldenCase]
-  -- ^ The test cases.
-  -> IO ()
-  -- ^ Run the tests.
-testValidity shouldSucceed contract invalids =
-  sequence_
-    [ assertBool (show (t, i)) $ (playTrace t contract i == o) == shouldSucceed
-    | (t, i, o) <- invalids
-    ]
 
 -- | A golden transaction.
 type GoldenTransaction = (State, Contract, TransactionInput, TransactionOutput)
