+++ marlowe-test/test/Spec/Marlowe/Plutus/Prelude.hs
@@ -1,15 +1,10 @@
------------------------------------------------------------------------------
+-- | Test the Plutus Prelude functions used by the Marlowe validators.
 --
 -- Module      :  $Headers
 -- License     :  Apache 2.0
 --
 -- Stability   :  Experimental
 -- Portability :  Portable
---
-
------------------------------------------------------------------------------
-
--- | Test the Plutus Prelude functions used by the Marlowe validators.
 module Spec.Marlowe.Plutus.Prelude (
   -- * Testing
   tests,
