+++ marlowe-test/test/Spec/Marlowe/Plutus/Script.hs
@@ -1,17 +1,14 @@
------------------------------------------------------------------------------
+{-# LANGUAGE NumericUnderscores #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE RecordWildCards #-}
+
+-- | Script execution functions for Marlowe testing.
 --
 -- Module      :  $Headers
 -- License     :  Apache 2.0
 --
 -- Stability   :  Experimental
 -- Portability :  Portable
---
------------------------------------------------------------------------------
-{-# LANGUAGE NumericUnderscores #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE RecordWildCards #-}
-
--- | Script execution functions for Marlowe testing.
 module Spec.Marlowe.Plutus.Script (
   -- * Evaluation
   evaluatePayout,
