+++ marlowe-test/test/Spec/Marlowe/Plutus/Transaction.hs
@@ -35,6 +35,7 @@ import Control.Monad (when)
 import Control.Monad.State (StateT, execStateT, lift)
 import Data.Bifunctor (bimap, second)
 import Data.List (nub)
+import Language.Marlowe.Core.V1.Merkle (MerkleizedContract (..), deepMerkleize, merkleizeInputs)
 import Language.Marlowe.Core.V1.Semantics (
   MarloweData (MarloweData),
   MarloweParams (..),
@@ -55,7 +56,7 @@ import Language.Marlowe.Core.V1.Semantics.Types (
   Token (Token),
   getInputContent,
  )
-import Language.Marlowe.Scripts (MarloweInput, MarloweTxInput (..))
+import Language.Marlowe.Scripts.Types (MarloweInput, MarloweTxInput (..))
 import Plutus.Script.Utils.Scripts (datumHash)
 import Plutus.V1.Ledger.Value (gt)
 import Plutus.V2.Ledger.Api (
@@ -107,9 +108,9 @@ import Spec.Marlowe.Plutus.Types (
 import Spec.Marlowe.Reference (ReferencePath, arbitraryReferenceTransaction)
 import Spec.Marlowe.Semantics.Arbitrary (arbitraryGoldenTransaction, arbitraryPositiveInteger)
 import Spec.Marlowe.Semantics.Golden (GoldenTransaction)
-import Spec.Marlowe.Semantics.Merkle (deepMerkleize, merkleizeInputs)
 import Test.Tasty.QuickCheck (Arbitrary (..), Gen, frequency, listOf, suchThat)
 
+import Control.Monad.Writer (runWriter)
 import qualified Language.Marlowe.Core.V1.Semantics.Types as M (Party (Address))
 import qualified Plutus.V1.Ledger.Value as V (adaSymbol, adaToken, singleton)
 import qualified PlutusTx.AssocMap as AM (fromList, toList)
@@ -449,7 +450,7 @@ validPayoutTransaction noisy =
     infoData <>= AM.fromList [inData]
     scriptPurpose .= Spending (txInInfoOutRef inScript)
 
-    -- The datum is the currency symbole and role name.
+    -- The datum is the currency symbol and role name.
     datum .= inDatum
 
     -- The redeemer is unit.
@@ -533,12 +534,12 @@ merkleize =
     contract <- use inputContract
     inputs <- use input
     -- Merkleize the contract and the input.
-    let (contract', continuations) = deepMerkleize contract
-        inputs' = maybe (error "Merkleization of inputs failed.") id $ merkleizeInputs continuations state contract' inputs
+    let (mcContract, mcContinuations) = runWriter $ deepMerkleize contract
+        inputs' = either error id $ merkleizeInputs MerkleizedContract{..} state inputs
     -- Update the contract, inputs, and outputs.
-    inputContract .= contract'
+    inputContract .= mcContract
     input .= inputs'
-    output .= computeTransaction inputs' state contract'
+    output .= computeTransaction inputs' state mcContract
 
 -- | Generate an arbitrary, valid Marlowe payout transaction: datum, redeemer, and script context.
 arbitraryPayoutTransaction
