+++ marlowe-test/test/Spec/Marlowe/Semantics.hs
@@ -20,9 +20,9 @@ import Test.Tasty (TestTree, testGroup)
 import qualified Spec.Marlowe.Semantics.Compute (tests)
 import qualified Spec.Marlowe.Semantics.Entropy (tests)
 import qualified Spec.Marlowe.Semantics.Functions (tests)
-import qualified Spec.Marlowe.Semantics.Golden (tests)
+import qualified Spec.Marlowe.Semantics.Golden.Tests (tests)
 import qualified Spec.Marlowe.Semantics.Next (tests)
-import qualified Spec.Marlowe.Semantics.Path (tests)
+import qualified Spec.Marlowe.Semantics.Path.Tests (tests)
 
 -- | Run the tests.
 tests :: TestTree
@@ -33,6 +33,6 @@ tests =
     , Spec.Marlowe.Semantics.Entropy.tests
     , Spec.Marlowe.Semantics.Functions.tests
     , Spec.Marlowe.Semantics.Compute.tests
-    , Spec.Marlowe.Semantics.Golden.tests
-    , Spec.Marlowe.Semantics.Path.tests
+    , Spec.Marlowe.Semantics.Golden.Tests.tests
+    , Spec.Marlowe.Semantics.Path.Tests.tests
     ]
