+++ marlowe-test/test/Spec/Marlowe/Semantics/Entropy.hs
