+++ marlowe-test/test/Spec/Marlowe/Semantics/Next/Common/Isomorphism.hs
