+++ marlowe-test/test/Spec/Marlowe/Semantics/Next/Common/QuickCheck.hs
