+++ marlowe-test/test/Spec/Marlowe/Semantics/Next/Contract/Generator.hs
