+++ marlowe-test/test/Spec/Marlowe/Semantics/Next/Contract/When/Choice.hs
