+++ marlowe-test/test/Spec/Marlowe/Serialization/CoreJson.hs
